fixed minor stuff with api
