--- ComicVine.API/Repository/IPostRepository.cs
@@ -1,9 +1,0 @@
-﻿using ComicVine.API.Controllers;
-using ComicVine.API.Models;
-
-namespace ComicVine.API.Repository;
-
-public interface IPostRepository
-{
-    public Task<PostPage> GetPostPage(string path, int pageNo, ILogger<PostController> logger);
-}
