--- ComicVine.API/Repository/IProfileRepository.cs
@@ -1,14 +1,0 @@
-﻿using ComicVine.API.Models;
-using ComicVine.API.Controllers;
-
-namespace ComicVine.API.Repository;
-
-public interface IUserRepository<T>
-{
-    public Task<Profile> GetProfile(string username, ILogger<T> logger);
-    public Task<FollowingPage> GetUserFollowing(string username, int pageNo, ILogger<T> logger);
-    public Task<FollowersPage> GetUserFollowers(string username, int pageNo, ILogger<T> logger);
-    public Task<BlogPage> GetUserBlog(string username, int pageNo, ILogger<T> logger);
-    public Task<ImagePage> GetUserImages(string username, int pageNo, ILogger<T> logger);
-    public Task GetUserForumPosts(string username);
-}
