+++ ComicVine.API/Repository/ProfileRepository.cs
@@ -6,6 +6,17 @@ using HtmlAgilityPack;
 
 namespace ComicVine.API.Repository;
 
+public interface IUserRepository<T>
+{
+    public Task<Profile> GetProfile(string username, ILogger<T> logger);
+    public Task<FollowingPage> GetUserFollowing(string username, int pageNo, ILogger<T> logger);
+    public Task<FollowersPage> GetUserFollowers(string username, int pageNo, ILogger<T> logger);
+    public Task<BlogPage> GetUserBlog(string username, int pageNo, ILogger<T> logger);
+    public Task<ImagePage> GetUserImages(string username, int pageNo, ILogger<T> logger);
+    public Task GetUserForumPosts(string username);
+    public Task<IEnumerable<Comicvine.Core.Parsers.Blog>> GetBlog(string username);
+}
+
 public class UserRepository: IUserRepository<ProfileController>
 {
    
@@ -64,6 +75,13 @@ public class UserRepository: IUserRepository<ProfileController>
         BlogPage blogPage = BlogParser.Parse(rootNode, pageNo, logger);
 
         return blogPage;
+    }
+
+    public async Task<IEnumerable<Comicvine.Core.Parsers.Blog>> GetBlog(string username) {
+        await using Stream stream = await Repository.GetStream($"/profile/{username}/blog");
+
+        HtmlNode rootNode = Repository.GetRootNode(stream);
+        return Comicvine.Core.Parsers.ParseBlog(rootNode);
     }
 
     public async Task<ImagePage> GetUserImages(string username, int pageNo, ILogger<ProfileController> logger) {
